+++ examples/apps/web-router/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lit-todomvc
 
+## 0.4.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.0.2
+
 ## 0.4.7
 
 ### Patch Changes